ai_backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Database
supabase==2.9.1

# AI APIs (Modular - easy to swap)
google-generativeai==0.8.3
# openai==1.54.0  # Optional
# anthropic==0.39.0  # Optional

# NLP - Korean
kiwipiepy==0.18.0

# Sentence Embeddings (for topic analysis)
sentence-transformers==3.2.0

# Caching
redis==5.2.0

# Utils
python-dotenv==1.0.1
httpx==0.27.2
apscheduler==3.10.4
pytest==8.3.3
python-json-logger==2.0.7
sentry-sdk==2.18.0
